2020/06/05 update.end search.not end

+++ DBTest/src/main/java/com/example/demo/ServiceImpl/UserInfoDeleteServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.example.demo.constants.message.ConstantsMsg;
 import com.example.demo.form.UserInfoDeleteForm;
 import com.example.demo.formDetail.UserInfoDeleteFormDetail;
+import com.example.demo.mapper.SalaryInfoMapper;
 import com.example.demo.mapper.UserInfoMapper;
 import com.example.demo.service.UserInfoDeleteService;
 import com.example.demo.utils.StringUtils;
@@ -18,36 +19,51 @@ public class UserInfoDeleteServiceImpl implements UserInfoDeleteService {
 	@Autowired
 	UserInfoMapper userInfoMapper;
 	@Autowired
+	SalaryInfoMapper salaryInfoMapper;
+	@Autowired
 	StringUtils stringUtils;
 
 	@Override
 	public List<UserInfoDeleteForm> userInfoSelectForDelete() {
 
-		List<UserInfoDeleteForm> userInfoList = userInfoMapper.selectIdName();
+		List<UserInfoDeleteForm> userInfoList = userInfoMapper.selectIdNameForDalete();
 
 		return userInfoList;
 	}
 
 	@Override
-	public UserInfoDeleteFormDetail userInfoDelete(String[] checkBox) {
+	public UserInfoDeleteFormDetail userInfoDelete(UserInfoDeleteFormDetail userInfoDeleteFormDetail) {
 
+		//メッセージの初期化
+		userInfoDeleteFormDetail.setMessage("");
+
+		//チェックボックスが選択されていない場合
+		if (userInfoDeleteFormDetail.getCheckBox() == null) {
+			userInfoDeleteFormDetail.setMessage(ConstantsMsg.MSG_NOT_SELECT);
+			return userInfoDeleteFormDetail;
+		}
+
+		//選択したチェックボックスの値をDELETE文に使用できるように整形
 		boolean result = true;
 		String idList = "";
-		for (String str : checkBox) {
+		for (String str : userInfoDeleteFormDetail.getCheckBox()) {
 			idList = idList + " " + str;
 		}
 		idList = stringUtils.trim(idList);
 		idList = idList.replaceAll(" ", ",");
-		UserInfoDeleteFormDetail userInfoDeleteFormDetail = new UserInfoDeleteFormDetail();
 
 		try {
+			//DELETE
 			userInfoMapper.deleteData(idList);
+			salaryInfoMapper.deleteData(idList);
 			userInfoDeleteFormDetail.setMessage(ConstantsMsg.MSG_DELETE_OK);
-			userInfoDeleteFormDetail.setUserInfoDeleteFormList(userInfoMapper.selectIdName());
+
+			//DELETE後のリストを取得
+			userInfoDeleteFormDetail.setUserInfoDeleteFormList(userInfoMapper.selectIdNameForDalete());
 
 		} catch (Exception e) {
 			userInfoDeleteFormDetail.setMessage(ConstantsMsg.MSG_DELETE_NG);
-			userInfoDeleteFormDetail.setUserInfoDeleteFormList(userInfoMapper.selectIdName());
+			userInfoDeleteFormDetail.setUserInfoDeleteFormList(userInfoMapper.selectIdNameForDalete());
 		}
 
 		return userInfoDeleteFormDetail;
